Crate sxd_xpath

Namespaces

The XPath specification assumes that the XML being processed has a prefix defined for every namespace present in the document. If you are processing XML that was parsed from text, this will be true by construction.

If you have programmatically created XML with namespaces but not defined prefixes, some XPath behavior may be confusing:

The name method will not include a prefix, even if the element or attribute has a namespace.

The namespace axis will not include namespaces without prefixes.

Reexports

pub use function::Function;

Modules

axis

function

nodeset

parser

tokenizer

Macros

nodeset!
Convenience constructor for a nodeset

Structs

EvaluationContext

Factory

Enums

LiteralValue

Value

Traits

Expression

Type Definitions

Functions

Namespaces

Variables