Trait sxd_xpath::axis::Axis

pub trait Axis: Debug {
    fn select_nodes<'a, 'd>(&self, context: &EvaluationContext<'a, 'd>, node_test: &NodeTest, result: &mut Nodeset<'d>);

    fn principal_node_type(&self) -> PrincipalNodeType { ... }
}

A directed traversal of Nodes.

Required Methods

fn select_nodes<'a, 'd>(&self, context: &EvaluationContext<'a, 'd>, node_test: &NodeTest, result: &mut Nodeset<'d>)

Applies the given node test to the nodes selected by this axis, adding matching nodes to the nodeset.

Provided Methods

fn principal_node_type(&self) -> PrincipalNodeType

Describes what node type is naturally selected by this axis.