Struct sxd_xpath::context::Evaluation

pub struct Evaluation<'c, 'd: 'c> {
    pub node: Node<'d>,
    pub position: usize,
    pub size: usize,
    /* private fields */
}

The context during evaluation of an XPath expression.

Clients of this library will use this when implementing custom functions.

Lifetimes

We track two separate lifetimes: that of the user-provided context ('c) and that of the document ('d). This allows the user-provided context to live shorter than the document.

Fields
node: Node<'d>

The context node

position: usize

The context position

size: usize

The context size

Implementations

impl<'c, 'd> Evaluation<'c, 'd>

pub fn new(context: &'c Context<'d>, node: Node<'d>) -> Evaluation<'c, 'd>

Prepares the context used while evaluating the XPath expression

pub fn new_context_for<N>(&self, node: N) -> Evaluation<'c, 'd>where N: Into<Node<'d>>,

Creates a new context node using the provided node

pub fn function_for_name(&self, name: QName<'_>) -> Option<&'c dyn Function>

Looks up the function with the given name

pub fn value_of(&self, name: QName<'_>) -> Option<&Value<'d>>

Looks up the value of the variable

pub fn namespace_for(&self, prefix: &str) -> Option<&str>

Looks up the namespace URI for the given prefix

pub fn new_contexts_for( self, nodes: OrderedNodes<'d> ) -> EvaluationNodesetIter<'c, 'd>

Yields a new Evaluation context for each node in the nodeset.