Enum sxd_xpath::nodeset::Node

pub enum Node<'d> {
    Root(Root<'d>),
    Element(Element<'d>),
    Attribute(Attribute<'d>),
    Text(Text<'d>),
    Comment(Comment<'d>),
    Namespace(Namespace<'d>),
    ProcessingInstruction(ProcessingInstruction<'d>),
}

Any of the various types of nodes found in an XML document.

Variants

Root(Root<'d>)

Element(Element<'d>)

Attribute(Attribute<'d>)

Text(Text<'d>)

Comment(Comment<'d>)

Namespace(Namespace<'d>)

ProcessingInstruction(ProcessingInstruction<'d>)

Implementations

impl<'d> Node<'d>

pub fn document(&self) -> Document<'d>

The document to which this node belongs.

pub fn prefixed_name(&self) -> Option<String>

The name of the node, including a prefix that corresponds to the namespace, if any.

pub fn expanded_name(&self) -> Option<QName<'d>>

Returns the expanded name of the node, if any.

pub fn parent(&self) -> Option<Node<'d>>

Returns the parent of the node, if any.

pub fn children(&self) -> Vec<Node<'d>>

Returns the children of the node, if any.

pub fn preceding_siblings(&self) -> Vec<Node<'d>>

Returns the nodes with the same parent that occur before this node.

pub fn following_siblings(&self) -> Vec<Node<'d>>

Returns the nodes with the same parent that occur after this node.

pub fn string_value(&self) -> String

Returns the string value of this node.

pub fn root(self) -> Option<Root<'d>>

pub fn element(self) -> Option<Element<'d>>

pub fn attribute(self) -> Option<Attribute<'d>>

pub fn text(self) -> Option<Text<'d>>

pub fn comment(self) -> Option<Comment<'d>>

pub fn processing_instruction(self) -> Option<ProcessingInstruction<'d>>

pub fn namespace(self) -> Option<Namespace<'d>>